python/flask-demo/requirements.txt
alembic==1.7.4
aniso8601==9.0.1
APScheduler==3.8.1
asn1crypto==1.3.0
attrs==21.2.0
backports.zoneinfo==0.2.1
bidict==0.21.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==8.0.3
coverage==6.0.2
cryptography==2.8
dnspython==2.1.0
eventlet==0.32.0
flasgger==0.9.5
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.4.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SocketIO==5.1.1
flask-sqlacodegen==1.1.8
Flask-SQLAlchemy==2.5.1
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==2.8
importlib-metadata==4.8.1
importlib-resources==5.3.0
inflect==5.3.0
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
Mako==1.1.5
MarkupSafe==2.0.1
mistune==0.8.4
netifaces==0.11.0
packaging==21.0
pluggy==1.0.0
portalocker==2.3.2
psutil==5.8.0
py==1.10.0
pycosat==0.6.3
pycparser==2.19
pycryptodome==3.15.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.5
pytest-cov==2.12.1
python-dateutil==2.8.2
python-engineio==4.3.0
python-socketio==5.4.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.22.0
ruamel-yaml==0.15.87
six==1.14.0
SQLAlchemy==1.4.40
toml==0.10.2
tqdm==4.42.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.25.8
Werkzeug==2.0.2
zipp==3.6.0
zope.event==4.5.0
zope.interface==5.4.0